requirements.txt
python-dotenv
openai
streamlit
Python-IO